requirements.txt
requests
google-generativeai
flask
TTS
